+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Fix Webtools event image handling [#178](https://github.com/rokwire/gateway-building-block/issues/178)
+
 ## [2.19.0] - 2026-01-26
 ### Changed
 - Webtools events not showing images [#175](https://github.com/rokwire/gateway-building-block/issues/175)